pyproject.toml
[project]
name = "minecrafttui"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "minecraft-launcher-lib==8.0",
    "pyrefly>=0.45.0",
    "python-dotenv>=1.2.2",
]
